requirements.txt
# Bwenge OS - Complete Backend Requirements

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis & Caching
redis==5.0.1
hiredis==2.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Data Validation
pydantic==2.5.0
email-validator==2.1.0

# File Processing
PyMuPDF==1.23.8  # PDF processing
python-docx==1.1.0  # DOCX processing
python-pptx==0.6.23  # PPTX processing
python-magic==0.4.27  # File type detection
aiofiles==23.2.1  # Async file operations

# AI & ML
openai==1.3.7
openai-whisper==20231117
tiktoken==0.5.2

# Vector Database
weaviate-client==3.25.3

# Task Queue
celery[redis]==5.3.4

# Payment Processing
stripe==7.8.0

# Rate Limiting
slowapi==0.1.9

# Monitoring & Observability
prometheus-client==0.19.0

# WebSocket
websockets==12.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
testcontainers==3.7.1

# Distributed Tracing
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-psycopg2==0.42b0

# Optional: For better performance
orjson==3.9.10  # Faster JSON serialization